requirements.txt
Django==2.0
gunicorn==19.7.1
djangorestframework==3.8.2
django-heroku
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-allauth==0.37.1
django-cors-headers==2.4.0